refactor: enum 파일들 도메인 엔티티로 이동
