+++ backend/build.gradle.kts
@@ -69,9 +69,9 @@ checkstyle {
     toolVersion = "10.12.5"  // 최신 안정 버전으로 업그레이드
     configFile = rootProject.file("config/checkstyle/naver-checkstyle-rules.xml")
     configProperties = mapOf(
-        "suppressionFile" to rootProject
-            .file("config/checkstyle/naver-checkstyle-suppressions.xml")
-            .absolutePath
+            "suppressionFile" to rootProject
+                    .file("config/checkstyle/naver-checkstyle-suppressions.xml")
+                    .absolutePath
     )
     isIgnoreFailures = false  // 명시적으로 설정
 }
@@ -91,6 +91,7 @@ val coverageExcludes = listOf(
         "**/dto/**",
         "**/exception/**",
         "**/response/**",
+        "**/repository/**",
         "**/error/**",
         "**/entity/**",
         "**/home/**",
